appsflyer/requirements.txt
datetime
requests
google-cloud-logging
google-cloud-bigquery
pandas
pyarrow